+++ common/config/models_main.go
@@ -5,6 +5,7 @@ type GeneralConfig struct {
 	Port             int    `yaml:"port"`
 	LogDirectory     string `yaml:"logDirectory"`
 	LogColors        bool   `yaml:"logColors"`
+	JsonLogs         bool   `yaml:"jsonLogs"`
 	TrustAnyForward  bool   `yaml:"trustAnyForwardedAddress"`
 	UseForwardedHost bool   `yaml:"useForwardedHost"`
 }